orchestrator/requirements.txt
anthropic>=0.18.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
pyarrow>=14.0.0       # required for pd.read_parquet / write_parquet
openpyxl>=3.1.0
pyyaml>=6.0
rich>=13.0.0
